Cross entity custom validation
