+++ eq-author-api/middleware/validateQuestionnaire.js
@@ -9,6 +9,7 @@ module.exports = logger => (req, res, next) => {
   const before = performance.now();
   req.validationErrors = validateQuestionnaire(req.questionnaire);
   const after = performance.now();
+  console.log(JSON.stringify(req.validationErrors));
   logger.info("Validation took %dms", after - before);
 
   return next();
