+++ eq-author-api/src/validation/customKeywords/index.js
@@ -1,0 +1,8 @@
+const previousAnswer = require("./previousAnswer");
+
+module.exports = ajv => {
+  return [previousAnswer].reduce(
+    (ajv, keyword) => ajv.addKeyword(keyword.keyword, keyword),
+    ajv
+  );
+};
